requirements.txt
# Crypto Regime Intelligence Engine - Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Processing
numpy==1.26.3
pandas==2.1.4

# Machine Learning
scikit-learn==1.4.0
joblib==1.3.2

# Deep Learning (for LSTM & CryptoBERT)
torch>=2.0.0
transformers>=4.30.0

# LLM Integration
openai>=1.0.0

# WebSocket
websockets>=12.0

# Utilities
python-dotenv==1.0.0
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Logging
structlog==24.1.0
